Move widget support to a separate plugin
