+++ package.json
@@ -46,6 +46,8 @@
     "watch:src": "tsc -w"
   },
   "dependencies": {
+    "@jupyter-widgets/base": "^3.0.0-alpha.2",
+    "@jupyter-widgets/jupyterlab-manager": "^3.0.0-alpha.2",
     "@jupyterlab/application": "^3.0.0-rc.5",
     "@jupyterlab/apputils": "^3.0.0-rc.5",
     "@jupyterlab/cells": "^3.0.0-rc.5",
@@ -54,8 +56,6 @@
     "@jupyterlab/filebrowser": "^3.0.0-rc.5",
     "@jupyterlab/notebook": "^3.0.0-rc.5",
     "@jupyterlab/ui-components": "^3.0.0-rc.5",
-    "@jupyter-widgets/base": "^3.0.0-alpha.2",
-    "@jupyter-widgets/jupyterlab-manager": "^3.0.0-alpha.2",
     "@lumino/coreutils": "^1.5.3",
     "@lumino/widgets": "^1.14.0",
     "@types/codemirror": "^0.0.97",
