+++ src/editor/format.ts
@@ -1,0 +1,13 @@
+export type DashboardInfo = {
+  version: number;
+  activeView: string;
+  views: { [id: string]: DashboardView };
+};
+
+export type DashboardView = {
+  name: string;
+  type: string;
+  cellMargin: number;
+  cellHeight: number;
+  numColumns: number;
+};
